spider/requirements.txt
pyquery==1.4.3
aiohttp==3.8.3
asyncio==3.4.3
python-decouple==3.6
pymongo==4.3.2
python-dateutil==2.8.2
thefuzz==0.19.0
python-levenshtein==0.20.8
numpy==1.23.4
python-daemon==2.3.2
psutil==5.9.4
slack_sdk==3.18.1